requirements.txt
Flask>=2.0.1
google-cloud-datastore>=2.1.6
python-dotenv
authlib
requests
python-jose
flask-cors
six
json2html

